# src/main/webapp/public/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "BillPagedFuelBill" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FuelBill"
            }
          }
        }
      },
...
      "Car" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "odometer" : {
            "type" : "integer",
            "format" : "int32"
          },
          "amountBills" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "CarInput" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "odometer" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "CarOverview" : {
        "type" : "object",
        "properties" : {
          "totalRefuels" : {
            "type" : "integer"
          },
          "totalMaintenanceEvents" : {
            "type" : "integer"
          },
          "totalMiscellaneousEvents" : {
            "type" : "integer"
          },
          "totalCost" : {
            "type" : "number"
          },
          "avgConsumption" : {
            "type" : "number"
          }
        }
      },
...
      "FuelBill" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "distance" : {
            "type" : "number"
          },
          "unit" : {
            "type" : "number"
          },
          "pricePerUnit" : {
            "type" : "number"
          },
          "estimateConsumption" : {
            "type" : "number"
          },
          "avgConsumption" : {
            "type" : "number"
          },
          "total" : {
            "type" : "number"
          }
        }
      },
      "FuelBillInput" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "number"
          },
          "vatRate" : {
            "type" : "integer"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "distance" : {
            "type" : "number"
          },
          "unit" : {
            "type" : "number"
          },
          "pricePerUnit" : {
            "type" : "number"
          },
          "estimateConsumption" : {
            "type" : "number"
          }
        }
      },
...
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
...
    "/api/bill/{carId}/fuel" : {
      "parameters" : [ {
        "name" : "carId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "put" : {
        "operationId" : "addNewFuelBill",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FuelBillInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Bill created."
          },
          "409" : {
            "description" : "Car is not for your user."
          },
          "400" : {
            "description" : "Validation failed, please check the response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : { }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong while saving. Please ask the server admin for help."
          }
        },
        "summary" : "Add New Bill",
        "tags" : [ "Fuel Resource" ]
      }
    },
    "/api/bill/{carId}/fuel/all" : {
      "parameters" : [ {
        "name" : "carId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "get" : {
        "description" : "Gets all bills for specified car, Pages starting at 1",
        "operationId" : "getAllFuelBills",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "year",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Bills found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BillPagedFuelBill"
                }
              }
            }
          }
        },
        "summary" : "Get All My Bills",
        "tags" : [ "Fuel Resource" ]
      }
    },
    "/api/bill/{carId}/fuel/years" : {
      "parameters" : [ {
        "name" : "carId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "get" : {
        "description" : "Gets all years of bills for specified car",
        "operationId" : "getAllFuelBillYears",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Bills found and years extracted.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "integer",
                    "format" : "int32"
                  }
                }
              }
            }
          }
        },
        "summary" : "Get All My Bills",
        "tags" : [ "Fuel Resource" ]
      }
    },
...
    "/api/car/my" : {
      "get" : {
        "operationId" : "getMyCars",
        "responses" : {
          "200" : {
            "description" : "Get all cars.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Car"
                  }
                }
              }
            }
          }
        },
        "summary" : "Get My Cars",
        "tags" : [ "Car Resource" ]
      },
      "put" : {
        "operationId" : "createCar",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CarInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Car created."
          },
          "400" : {
            "description" : "Maximal amount of cars created or input was invalid."
          },
          "500" : {
            "description" : "Something went wrong while saving. Please ask the server admin for help."
          }
        },
        "summary" : "Create Car",
        "tags" : [ "Car Resource" ]
      }
    },
    "/api/car/my/{id}" : {
      "post" : {
        "operationId" : "updateMyCar",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CarInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Car found and updated."
          },
          "400" : {
            "description" : "Input was invalid."
          }
        },
        "summary" : "Update Car",
        "tags" : [ "Car Resource" ]
      },
      "get" : {
        "operationId" : "getMyCar",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Car found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Car"
                }
              }
            }
          },
          "204" : {
            "description" : "Id was not found."
          }
        },
        "summary" : "Get My Car",
        "tags" : [ "Car Resource" ]
      }
    },
    "/api/car/my/{id}/overview" : {
      "get" : {
        "operationId" : "getMyCarOverview",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Car found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CarOverview"
                }
              }
            }
          },
          "204" : {
            "description" : "Id was not found."
          }
        },
        "summary" : "Get My Car Overview",
        "tags" : [ "Car Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "car-ledger API",
    "version" : "0.1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}